Add "WebViewportItem" class, the "scroller" interaction item
This QGraphicsItem is supposed to catch user interaction (panning and
zooming gestures) and make QGraphicsWebView to present these
interactions to the user.

Currently only proof of concept. Panning == follow mouse move when
button is pressed. Zooming == scrollwheel OUTSIDE web area.

Also, the structure of the code is a mess due to old code.