Panning: Move panned widget scaling anim to WebViewport
This is needed so that we can conform to MPannableViewport API.

Also otherwise good practice. Originally this was due to not being
able to detect panned item resizes. This is still the case,
and the commit causes regression of not showing toolbars when
resize happens. This should be fixed later.